Condense compnent files
